django_intro/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.3
django-allauth==0.61.1
django-bootstrap5==23.4
django-cachalot==2.6.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-extra-views==0.14.0
django-filter==24.1
django-fontawesome-6==1.0.0.0
django-jsonform==2.22.0
django-model-utils==4.4.0
django-pwa==1.1.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-stronghold==0.4.0
django-tables2==2.7.0
djangorestframework==3.15.0
filelock==3.13.1
flake8==7.0.0
humanize==4.9.0
identify==2.5.35
idna==3.6
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
platformdirs==4.2.0
pre-commit==3.6.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.4.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
